chore: api 함수를 response 종류별로 분류 및 함수별 설명 추가
