+++ src/apis/notices.ts
@@ -1,3 +1,4 @@
+import { GetShopNoticesResponse } from '../types/apis/notice/getShopNotices';
 import type { GetShopsSingleNoticeResponse } from '../types/apis/notice/getShopSingleNotice';
 import {
   GetNoticesRequest,
@@ -19,6 +20,26 @@ export const postNotice = async (request: PostNoticeRequest): Promise<PostNotice
   return response.data;
 };
 
+export const getNotice = async (request: GetNoticesRequest): Promise<GetNoticesResponse> => {
+  const response = await axiosInstance.get('/notices', { params: request });
+  return response.data;
+};
+
+export type { GetNoticesRequest };
+
+// 가게의 공고 목록 조회(의진)
+export const getShopNotices = async (
+  id: string,
+  page: number,
+  countPerPage: number
+): Promise<GetShopNoticesResponse> => {
+  const response = await axiosInstance.get(
+    `/shops/${id}/notices?offset=${(page - 1) * countPerPage}&limit=${countPerPage}`
+  );
+  return response.data;
+};
+
+// 가게의 특정 공고 조회(의진)
 export const getShopSingleNotice = async (
   shopId: string,
   noticeId: string
@@ -26,10 +47,3 @@ export const getShopSingleNotice = async (
   const response = await axiosInstance.get(`/shops/${shopId}/notices/${noticeId}`);
   return response.data;
 };
-
-export const getNotice = async (request: GetNoticesRequest): Promise<GetNoticesResponse> => {
-  const response = await axiosInstance.get(`/notices`, { params: request });
-  return response.data;
-};
-
-export type { GetNoticesRequest };
